ADIL_RAG_Chatbot/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
openai>=1.0.0

numpy
